+++ server/src/main/java/de/ollie/carp/bm/rest/v1/dto/BattleMapTokenDTO.java
@@ -1,5 +1,6 @@
 package de.ollie.carp.bm.rest.v1.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 import lombok.Data;
 import lombok.Generated;
@@ -13,6 +14,6 @@ public class BattleMapTokenDTO {
 	private UUID id;
 	private BattleMapDTO battleMap;
 	private TokenDTO token;
-	private double fieldX;
-	private double fieldY;
+	private BigDecimal fieldX;
+	private BigDecimal fieldY;
 }
