+++ service-interface/src/main/java/de/ollie/carp/bm/core/model/BattleMapToken.java
@@ -1,5 +1,6 @@
 package de.ollie.carp.bm.core.model;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 import lombok.Data;
 import lombok.Generated;
@@ -13,6 +14,6 @@ public class BattleMapToken {
 	private UUID id;
 	private BattleMap battleMap;
 	private Token token;
-	private double fieldX;
-	private double fieldY;
+	private BigDecimal fieldX;
+	private BigDecimal fieldY;
 }
